requirements.txt
flask
flask-bootstrap
requests
sqlalchemy
mysqlclient
numpy
pygdal
Pillow